+++ database/helpers/database_manager.py
@@ -2,17 +2,17 @@ import time
 
 from sqlalchemy import create_engine, inspect, text
 from sqlalchemy.exc import OperationalError
+from sqlalchemy.orm import sessionmaker
 
-from config import get_database_url
-from models import Base, SessionLocal
+from database.config import get_database_url
 
 
 class DatabaseManager:
     def __init__(self):
         self.database_url = get_database_url()
         self.engine = self._create_database_engine()
-        SessionLocal.configure(bind=self.engine)
-        self.session = SessionLocal()
+        self.SessionLocal = sessionmaker(bind=self.engine)
+        self.session = self.SessionLocal()
 
     def _create_database_engine(self, echo=True, max_retries=3, retry_delay=2):
         """Create and return a database engine with retry logic"""
@@ -36,23 +36,3 @@ class DatabaseManager:
 
     def get_session(self):
         return self.session
-
-    def create_tables(self):
-        try:
-            inspector = inspect(self.engine)
-            existing_tables = inspector.get_table_names()
-            all_table_names = [table.name for table in Base.metadata.sorted_tables]
-            tables_to_create = [
-                table for table in all_table_names if table not in existing_tables
-            ]
-
-            if tables_to_create:
-                print(f"Creating tables: {tables_to_create}")
-                Base.metadata.create_all(self.engine)
-                print("Tables created successfully!")
-            else:
-                print("All tables already exist")
-            return True
-        except Exception as e:
-            print(f"Error creating tables: {e}")
-            raise
